Add files via upload
twt6 file produces: val.png, which is a graph generated from analysis of the validation dataset. the datasets are twitter_training and twitter_validation.

twt7 file produces: val1.png. i added column names and modified the aforementioned datasets in wps office and saved it as twitter_training1 and twitter_validation1.